+++ src/components/auth/AuthClient.tsx
@@ -3,39 +3,7 @@ import 'src/styles/AuthStyles.css';
 import { Authenticator, ThemeProvider, useTheme } from '@aws-amplify/ui-react';
 import { ComponentOverrides, FormFieldsOverrides } from './AuthOverrides';
 import { signUp, type SignUpInput } from 'aws-amplify/auth';
-import { Source_Sans_3 } from 'next/font/google';
-
-const sourceSans3 = Source_Sans_3({
-  subsets: ['latin'],
-  display: 'swap',
-});
-
-const loginTheme = {
-  name: 'login-theme',
-  tokens: {
-    colors: {
-      background: { primary: '#032436' },
-      border: {
-        primary: '#2B4A5C',
-        secondary: '#2B4A5C',
-      },
-      font: {
-        primary: '#FDFEFF',
-        hover: '#FDFEFF',
-        secondary: '#B1B5B8',
-      },
-    },
-
-    components: {
-      input: {},
-      authenticator: {
-        router: { borderWidth: '0', boxShadow: 'none' },
-        form: { padding: '0px' },
-        field: {},
-      },
-    },
-  },
-};
+import { loginTheme, sourceSans3 } from './AuthTheme';
 
 const AuthClient = () => {
   const { tokens } = useTheme();
@@ -65,7 +33,9 @@ const AuthClient = () => {
   };
 
   return (
-    <div className={`bg-cardsDark ${sourceSans3.className} rounded-10 p-8 `}>
+    <div
+      className={`bg-cardsDark ${sourceSans3.className} rounded-10 p-8 shadow-sm `}
+    >
       <ThemeProvider theme={loginTheme}>
         <div className="mb-12">
           <p className="text-primary font-extralight leading-8">Welcome to</p>
